pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blogplatform"
version = "0.1.0"
description = "A Django-based blog platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.7",
    "django-allauth>=65.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "requests>=2.32.5"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["blogplatform*", "articles*", "comments*", "likes*", "common*"]
exclude = ["tests*"]
